refactor: 모달 컴포넌트 추출에 따른 변경
