+++ src/components/modal/RestaurantDetailModal.ts
@@ -1,8 +1,8 @@
 import Modal from './Modal';
 import RestaurantRepository from '../../domain/RestaurantRepository';
 import { $addEvent } from '../../utils/dom';
-import favoriteFilledIcon from '../../assets/favorite-icon-filled.png';
-import favoriteLinedIcon from '../../assets/favorite-icon-lined.png';
+import favoriteFilledIcon from '../assets/favorite-icon-filled.png';
+import favoriteLinedIcon from '../assets/favorite-icon-lined.png';
 
 class RestaurantDetailModal extends Modal {
   static observedAttributes: string[] = ['key', 'open'];
@@ -20,13 +20,25 @@ class RestaurantDetailModal extends Modal {
   }
 
   setEvent(): void {
-    $addEvent(this, '.button--primary', 'click', () => this.updateModal(false));
-    $addEvent(this, '.button--secondary', 'click', () => this.#removeRestaurant());
+    $addEvent(this, '.button--primary', 'click', this.#closeModal.bind(this));
+    $addEvent(this, '.button--secondary', 'click', this.#removeRestaurant.bind(this));
+  }
+
+  #closeModal(): void {
+    this.updateModal(false);
+    this.makeCustomEvent('updateRestaurantList');
   }
 
   #removeRestaurant(): void {
     RestaurantRepository.removeRestaurant(this.#key);
     this.makeCustomEvent('updateRestaurantList');
+  }
+
+  #toggleFavorite(): void {
+    RestaurantRepository.toggleFavoriteRestaurant(this.#key);
+
+    this.#restaurant = RestaurantRepository.getRestaurant(this.#key);
+    this.render();
   }
 
   #handleFavoriteIcon(isFavorite: boolean): string {
@@ -38,14 +50,14 @@ class RestaurantDetailModal extends Modal {
   modalContent(): string {
     return `
         <div class="restaurant-detail-container">
-        <button class="favorite__button">
+          <button class="favorite__button">
             ${this.#handleFavoriteIcon(this.#restaurant.isFavorite)}
-        </button>
-        <category-icon category=${this.#restaurant.category}></category-icon>
-        <h2 class="restaurant__name text-subtitle">${this.#restaurant.name}</h2>
-        <span class="restaurant__distance text-body">캠퍼스부터 ${this.#restaurant.distance}분 내</span>
-        <p class="restaurant__description text-body">${this.#restaurant.description || ''}</p>
-        <p class="restaurant__reference text-body">${this.#restaurant.reference || ''}</p>
+          </button>
+          <category-icon category=${this.#restaurant.category}></category-icon>
+          <h2 class="restaurant__name text-subtitle">${this.#restaurant.name}</h2>
+          <span class="restaurant__distance text-body">캠퍼스부터 ${this.#restaurant.distance}분 내</span>
+          <p class="restaurant__description text-body">${this.#restaurant.description || ''}</p>
+          <p class="restaurant__reference text-body">${this.#restaurant.reference || ''}</p>
         </div>
         <div class="button-container">
             <button type="button" class="button button--secondary text-caption">삭제하기</button>
